backend/requirements.txt
aiosqlite==0.20.0
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.6.0
python-dotenv==1.0.1
google-genai==0.3.0
httpx==0.27.2
icalendar==6.0.1
python-multipart==0.0.12
